Added empty text when list is empty
